+++ packages/builder/src/pages/builder/app/[application]/design/[screenId]/_components/BindingNode.svelte
